[skip travis] Move unit test to classifier package
